Moved JSON store construction to IO.java for API consistency, update ImportControler with store call, added url.
